+++ src/components/mobile/RenderReward.tsx
@@ -7,15 +7,20 @@ import {
   numberWithCommas,
   numberWithCommasToFixed,
 } from "../../utilities/TextFormatter";
-import { formatMoney } from "../../utilities/TextFormatter";
-
+import { DateTimeUtil } from "../../utilities/DateTimeUtil";
+import { convertBuddhistYear } from "../../utilities/ConvertToBuddhistYear";
+import moment from "moment";
+import color from "../../resource/color";
 interface RenderReward {
   img: string;
   name: any;
   description: any;
   condition: any;
   point: any;
-  count: any;
+  type: string | null;
+  endUseDateTime: any;
+  exChange: string | null;
+  countdownTime: any;
 }
 
 const RenderReward: React.FC<RenderReward> = ({
@@ -24,7 +29,10 @@ const RenderReward: React.FC<RenderReward> = ({
   description,
   condition,
   point,
-  count,
+  type,
+  endUseDateTime,
+  exChange,
+  countdownTime,
 }) => {
   const div = useRef<any>();
   useLayoutEffect(() => {
@@ -40,7 +48,6 @@ const RenderReward: React.FC<RenderReward> = ({
     window.addEventListener("scroll", onScroll);
     return () => window.removeEventListener("scroll", onScroll);
   }, []);
-
   return (
     <div className="col-lg-12">
       <div ref={div}>
@@ -60,41 +67,75 @@ const RenderReward: React.FC<RenderReward> = ({
               height: "100%",
             }}
           >
-            <Image
-              src={img ? img : uploadImg}
-              width={"100%"}
-              height={"26.7%"}
-              preview={false}
-            />
+            <div className="p-3">
+              <Image
+                style={{ borderRadius: 10 }}
+                src={img ? img : uploadImg}
+                width={"100%"}
+                height={"26.7%"}
+                preview={false}
+              />
+            </div>
+
             <div className="p-4">
-              <p style={{ fontSize: "23px", fontWeight: "600" }}>
+              <h5 style={{ fontWeight: "600" }}>
                 {!name ? "ชื่อของรางวัล" : name}
-              </p>
-              <p style={{ fontSize: "18px", fontWeight: "400" }}>{`ใช้คะแนน  ${
-                !point ? 0 : numberWithCommas(parseFloat(point))
-              } คะแนน`}</p>
-              <div
-                style={{
-                  width: "170px",
-                  height: "62px",
-                  backgroundColor: "#FFEFDD",
-                  borderRadius: 12,
-                }}
-              >
-                <div style={{ paddingLeft: "12px", paddingTop: "8px" }}>
-                  <h1 style={{ color: "#B26003", fontSize: "15px" }}>
-                    จำนวนคงเหลือ
-                  </h1>
-                  <p style={{ fontSize: "17px", fontWeight: "bold" }}>
-                    {count}
-                  </p>
+              </h5>
+
+              {exChange === "MISSION" ? (
+                <p style={{ fontSize: "18px", fontWeight: "500" }}>ภารกิจ</p>
+              ) : (
+                <p style={{ fontSize: "18px", fontWeight: "500" }}>{`ใช้แต้ม  ${
+                  !point ? 0 : numberWithCommas(parseFloat(point))
+                } แต้ม`}</p>
+              )}
+              {type === "DIGITAL" && (
+                <div
+                  style={{
+                    width: "185px",
+                    height: "65px",
+                    backgroundColor: "rgba(220, 241, 254, 1)",
+                    borderRadius: 10,
+                    padding: 10,
+                  }}
+                >
+                  <div
+                    style={{
+                      color: "rgba(12, 100, 141, 1)",
+                      fontSize: "15spanx",
+                      fontWeight: "500",
+                      lineHeight: "22px",
+                      paddingLeft: "12px",
+                    }}
+                  >
+                    {`หมดอายุอีก ${countdownTime ? countdownTime : "0"} วัน`}
+                  </div>
+                  <div
+                    style={{
+                      color: "black",
+                      fontSize: "18px",
+                      fontWeight: "700",
+                      lineHeight: "22px",
+                      paddingLeft: "12px",
+                    }}
+                  >
+                    {endUseDateTime
+                      ? convertBuddhistYear.toBuddhistYear(
+                          moment(endUseDateTime),
+                          "DD MMM YY"
+                        )
+                      : "-"}
+                  </div>
                 </div>
-              </div>
-              <br />
-              <p>รายละเอียด</p>
-              <p>{!description ? "-" : parse(description)}</p>
+              )}
+              <p className="pt-3">รายละเอียด</p>
+              <p style={{ fontWeight: "400", color: "rgba(107, 117, 128, 1)" }}>
+                {!description ? "-" : parse(description)}
+              </p>
               <p>เงื่อนไข</p>
-              <p>{!condition ? "-" : parse(condition)}</p>
+              <p style={{ fontWeight: "400", color: "rgba(107, 117, 128, 1)" }}>
+                {!condition ? "-" : parse(condition)}
+              </p>
             </div>
           </div>
         </div>